+++ app/scripts/runtime.ts
@@ -1,9 +1,0 @@
-let EventEmitter = require('super-event-emitter');
-
-// Współdzielony obiekt między stanami.
-let runtime = Object.create(null);
-
-// Rozszerzamy obiekt gry o zdarzenia, aby np. podłączyć statystyki.
-EventEmitter.mixin(runtime);
-
-export default runtime;