+++ app/states/fly-states/avoid-monsters-state.js
@@ -1,0 +1,8 @@
+
+
+export default class AvoidMonstersState extends Phaser.State {
+
+    create() {
+
+    }
+}