+++ webpack.config.js
@@ -1,51 +1,28 @@
 'use strict';
 
-let WriteFilePlugin = require('write-file-webpack-plugin');
-
 module.exports = {
-    resolve: {
-        extensions: ['.ts', '.js', '']
-    },
-
-    devServer: {
-        outputPath: './app/dist'
-    },
-
     entry: {
-        bundle: './app/scripts/main',
-        phaser: './node_modules/phaser/build/phaser'
+        bundle: './app/main'
     },
 
     output: {
         filename: '[name].js',
-        path: './app/dist',
-        // pathinfo: true
+        path: './public/dist',
+        pathinfo: true
     },
 
-    // devtool: '#inline-source-map',
+    // devtool: '#source-map',
 
     module: {
         loaders: [
             {
-                test: /phaser\.js$/,
-                loader: 'script-loader'
-            },
-            {
                 test: /\.js$/,
                 loader: 'babel-loader'
-            },
-            {
-                test: /\.ts$/,
-                loader: 'ts-loader'
             },
             {
                 test: /\.json$/,
                 loader: 'json-loader'
             }
         ]
-    },
-
-    plugins: [
-        new WriteFilePlugin()
-    ]
+    }
 };